extract npm command to helper util
